refactor(FE): `NotificationBox` 컴포넌트의 시간 계산 함수 리팩토링
- `claculateTime` 함수를 `utils.ts`로 이동하여 코드 중복 제거
- `NotificationBox.tsx`에서 `claculateTime` 함수 임포트
